Move CompositeReadIO into namespace
Just moves the CompositeReadIO out into the Multipart::Post namespace.
